+++ mem/pmem.h
@@ -13,7 +13,7 @@
 
 typedef struct page {
   phys_addr_t phys_addr;
-  int order;
+  uint_least8_t order;
   struct page *next_free;
 } page_t;
 
@@ -34,7 +34,7 @@ void freelist_push(page_t **freelist, page_t *page);
 page_t *freelist_pop(page_t **freelist);
 page_t *pmem_get_page(phys_addr_t phys_addr);
 pmem_segment_t *pmem_get_seg(phys_addr_t phys_addr);
-void pmem_free(page_t *page, int order);
-page_t *pmem_alloc(int order);
+void pmem_free(page_t *page, uint_fast8_t order);
+page_t *pmem_alloc(uint_fast8_t order);
 
 #endif /* _MEM_PMEM_H */
